SPT-1969: Renamed TokenFunction
Renamed TokenFunction to PostTokenHandler, which now follows our
preferred naming strategy.
